Part 2 performance loading
Very basic part 2 performance load implementation.
Performane > Load Part 2 loads the performance into to TG 9-16 and adds 8 to each TG channel.
This is just a test of how it sounds this way.
In this implementation part 2 uses the same reverb chain as part 1, so it doesn't sound very good.

ToneGenerators=16 should be added to the minidexed.ini so you need rpi4 or rpi5.